ELEC-622: Create example folder
