+++ projects/fw_tutorial/src/potentiometer.c
@@ -5,5 +5,6 @@
 
 StatusCode potentiometer_init(const PotentiometerSettings *settings,
                               PotentiometerStorage *storage) {
+  // TODO(ELEC-624): Implement this once potentiometer pin has been re-mapped on hardware
   return STATUS_CODE_UNIMPLEMENTED;
 }
